+++ CHANGELOG.md
@@ -23,6 +23,10 @@ This project follows **Semantic Versioning (SemVer)** ([semver.org](https://semv
 
 - add sbom-aggregation directory to dependabot pip config
 
+### Fixed
+
+- address sonarcloud issues in `dependabot-configuration-helper.sh`
+
 ## [0.98.1] - 2026-03-17
 
 ### Added